fixed errors

+++ Portmonetka/Controllers/TransactionController.cs
@@ -22,7 +22,9 @@ namespace Portmonetka.Controllers
             if (_dbContext.Transactions == null)
                 return NotFound();
 
-            return await _dbContext.Transactions.ToListAsync();
+            return await _dbContext.Transactions
+                .Where(t => t.DateDeleted == null)
+                .ToListAsync();
         }
 
         [Route("id/{id}")]
@@ -46,7 +48,9 @@ namespace Portmonetka.Controllers
             if (_dbContext.Transactions == null)
                 return NotFound();
 
-            return await _dbContext.Transactions.Where(x => x.WalletId == walletId).ToListAsync();
+            return await _dbContext.Transactions
+                .Where(t => t.WalletId == walletId && t.DateDeleted == null)
+                .ToListAsync();
         }
 
         [HttpPost]
@@ -85,11 +89,55 @@ namespace Portmonetka.Controllers
             if (_dbContext.Transactions == null)
                 return NotFound();
 
-            var transaction = await (_dbContext.Transactions.FindAsync(id));
+            var transaction = await _dbContext.Transactions.FindAsync(id);
+
             if (transaction == null)
                 return NotFound();
 
-            _dbContext.Transactions.Remove(transaction);
+            if (typeof(Auditable).IsAssignableFrom(typeof(Transaction)))
+            {
+                transaction.DateDeleted = DateTimeOffset.UtcNow;
+                _dbContext.Transactions.Attach(transaction);
+                _dbContext.Entry(transaction).State = EntityState.Modified;
+            }
+            else
+            {
+                _dbContext.Transactions.Remove(transaction);
+            }
+
+            //_dbContext.Transactions.Remove(transaction);
+
+            await _dbContext.SaveChangesAsync();
+
+            return Ok();
+        }
+
+        [Route("delete")]
+        [HttpDelete]
+        public async Task<ActionResult> DeleteTransactiosn(IEnumerable<int> ids)
+        {
+            if (_dbContext.Transactions == null)
+                return NotFound();
+
+            var transactionsToDelete = _dbContext.Transactions.Where(t => ids.Any(id => id == t.Id));
+
+            if (transactionsToDelete == null)
+                return NotFound();
+
+            foreach ( var transaction in transactionsToDelete)
+            {
+                if (typeof(Auditable).IsAssignableFrom(typeof(Transaction)))
+                {
+                    transaction.DateDeleted = DateTimeOffset.UtcNow;
+                    _dbContext.Transactions.Attach(transaction);
+                    _dbContext.Entry(transaction).State = EntityState.Modified;
+                }
+                else
+                {
+                    _dbContext.Transactions.Remove(transaction);
+                }
+            }
+
             await _dbContext.SaveChangesAsync();
 
             return Ok();
@@ -102,7 +150,8 @@ namespace Portmonetka.Controllers
             if (_dbContext.Transactions == null)
                 return NotFound();
 
-            var transactions = _dbContext.Transactions.Where(t => t.WalletId == walletId);
+            var transactions = _dbContext.Transactions
+                .Where(t => t.WalletId == walletId);
 
             if (transactions != null)
                 _dbContext.Transactions.RemoveRange(transactions);
